moving necessary files from the `presentation` module to the `features:pws:presentation` module
