+++ features/pws/presentation/src/main/java/com/fredprojects/features/pws/presentation/PWListScreen.kt
@@ -1,4 +1,4 @@
-package com.fredprojects.helloworld.presentation.features.pws
+package com.fredprojects.features.pws.presentation
 
 import androidx.compose.animation.*
 import androidx.compose.foundation.clickable
@@ -14,11 +14,10 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.fredprojects.helloworld.presentation.R
-import com.fredprojects.helloworld.presentation.core.FredFloatingActionButton
-import com.fredprojects.helloworld.presentation.core.FredTopBar
-import com.fredprojects.helloworld.presentation.features.pws.vm.PWEvents
-import com.fredprojects.helloworld.presentation.features.pws.vm.PWState
+import com.fredprojects.core.ui.*
+import com.fredprojects.core.ui.R
+import com.fredprojects.features.pws.presentation.vm.PWEvents
+import com.fredprojects.features.pws.presentation.vm.PWState
 import kotlinx.coroutines.launch
 
 @Composable
