+++ features/pws/presentation/src/main/java/com/fredprojects/features/pws/presentation/vm/PWEvents.kt
@@ -1,7 +1,7 @@
-package com.fredprojects.helloworld.presentation.features.pws.vm
+package com.fredprojects.features.pws.presentation.vm
 
-import com.fredprojects.helloworld.domain.features.pws.models.PracticalWork
-import com.fredprojects.helloworld.domain.features.pws.utils.SortingPW
+import com.fredprojects.features.pws.domain.models.PracticalWork
+import com.fredprojects.features.pws.domain.utils.SortingPW
 
 /**
  * PWEvents is used to send events to the PWListVM to perform actions
